requirements.txt
python-dotenv
typing-extensions
langchain-openai
langgraph
streamlit
pandas
python-docx